Fix: Broken InputFile tests
Mocked the utility function blocking the test
